requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

-e .



colored==1.4.4

openai==0.27.6

packaging==23.1; python_version >= '3.7'

pygments==2.15.1; python_version >= '3.7'

python-dotenv==1.0.0